fixed level loading bug
Level loading bug was due to a quark with Phaser3.  Essentially I attempted to replace the current scenes tilemap using data provided in init(..) but Phaser3 doesn't allow us to simply overwrite and existing tilemap.  To fix this I made PlatformerScene a base class and updated the constructor/preload to provide a new tilemap and scene for each level.