Add `Player.normalBody` and `Player.crouchedBody` to modify collision bounds
We plan to use the Player crouch ability to provide access to hidden areas.  The way we restrict access is by using a small (single tile) entrance that is too large to normally enter but when the player is crouched the reduced collider size (1/2 of the original height) is small enough to allow entrance.